Add routing and tabs for Clients and Plan Details sections in Longview Landing
