+++ packages/manager/src/features/Longview/LongviewLanding/LongviewPlans.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export const LongviewPlans: React.FC<{}> = props => {
+  return <div>Hello world</div>;
+};
+
+export default LongviewPlans;
